+++ lang/en/messages.php
@@ -5,5 +5,8 @@ declare(strict_types=1);
 return [
     'close' => 'Close',
     'dismiss' => 'Dismiss',
+    'grid_view' => 'Grid view',
+    'list_view' => 'List view',
     'no_results' => 'No data found',
+    'view' => 'View',
 ];